Adds Ruby 3.2 to the CI matrix.  Requires Ruby >= 2.6. (#48)
Added Ruby 3.2 to the CI matrix. Also upgraded the checkout actions.

Ruby 2.5 no longer passes because it can't get versions of rubocop and rubocop-discourse that allow a clean Rubocop run. So, given that Ruby 2.5 was removed about a year ago, I chose to resolve this issue by removing Ruby 2.5.

---

Additionally updated JRuby versions in CI to match the 2.6 requirement.

Co-authored-by: Jarek Radosz <jradosz@gmail.com>